Replace Circle example by rvo_wrapper.cpp, with ROS and OpenMP
- Enabled OpenMP compiler flag in CMakeLists, which makes the simulation
much faster when using many agents. Will use this for running many
simulations in parallel with "omp parallel for"
