+++ icrin/rvo_wrapper/include/rvo_wrapper/Definitions.h
@@ -74,6 +74,14 @@
  */
 const float RVO_EPSILON = 0.00001f;
 
+// Console printing MACROS
+#define ERR(x) std::cerr << "\033[22;31;1m" << x << "\033[0m";    // RED
+#define WARN(x) std::cerr << "\033[22;33;1m" << x << "\033[0m";   // YELLOW
+#define INFO(x) std::cerr << "\033[22;37;1m" << x << "\033[0m";   // WHITE
+#define DEBUG(x) std::cerr << "\033[22;34;1m" << x << "\033[0m";  // BLUE
+#define CLEAR() std::cerr << "\x1B[2J\x1B[H";
+// CSI[2J clears screen, CSI[H moves the cursor to top-left corner
+
 namespace RVO {
 class Agent;
 class Obstacle;
